backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy

python-jose[cryptography]
passlib[bcrypt]

httpx
python-dotenv
pdfplumber
